All pulsed sweeps are written and working
Code still needs to be cleaned and commented a bit before merging two
branches...  Notes:

DC internal sweeps are significantly faster than external sweeps where
you tell the SMU to source, trigger, measure, and return data
incrementally.

Pulsed internal sweeps are similarly fast, but the waveform isn't the
same as what it is for single triggers; this mode is probably suitable
for long (>500 us) pulses, where the rise time isn't as much of an
issue.

Pulsed external sweeps are slower, but the waveform peaks much more
quickly than the internal sweep.  This is more accurate.

The SMU is doing something odd with the signal from the photodiode; it
appears to peak, and then drop as the pulse continues.  The SRS transamp
can't handle the amount of current that would come from the devices, and
Yufei's transamp seems to be too slow and results in ringing as the
pulse reaches its end.